requirements.txt
yfinance
pandas-ta
plotly
tqdm
scipy
ta
tapy
datetime
keras
tensorflow
backtesting
pandas
numpy
seaborn
matplotlib
scikit-learn
Flask
gunicorn
flask-cors
alpaca-trade-api
alpaca-py
python-dotenv
apscheduler
dill
#-e .